refactor(application-config): move getCustomApplicationConfigPath to application-config package
