+++ packages/application-config/src/index.ts
@@ -1,4 +1,4 @@
 export { default as processConfig } from './process-config';
-export { explorer } from './load-config';
+export { getCustomApplicationConfigPath } from './load-config';
 export * from './errors';
 export * from './types';
